move error types to inside participant
