+++ src/participant/participantErrorTypes.ts
